style: changed font in SectionTitle and AboutMe to be JakartaSans
